Solucionando detalles de guardado de nodo

+++ backend/src/models.py
@@ -44,5 +44,5 @@ class Nodo(Base):
     __tablename__ = "nodo"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True, autoincrement=True)
-    posicionx : Mapped[float] = mapped_column(Integer, index=True, nullable=False)
-    posiciony : Mapped[float] = mapped_column(Integer, index=True, nullable=False)
+    posicionx : Mapped[float] = mapped_column(Float, index=True, nullable=False)
+    posiciony : Mapped[float] = mapped_column(Float, index=True, nullable=False)
